requirements.txt
pygame==2.0.1
schedule==1.1.0
pyserial==3.5